--- a/logbook-json/src/main/java/org/zalando/logbook/json/JsonMediaType.java
+++ b/logbook-json/src/main/java/org/zalando/logbook/json/JsonMediaType.java
@@ -1,15 +1,36 @@
 package org.zalando.logbook.json;
 
-import org.zalando.logbook.common.MediaTypeQuery;
-
 import java.util.function.Predicate;
 
 final class JsonMediaType {
 
     private JsonMediaType() {
-
     }
 
-    static final Predicate<String> JSON = MediaTypeQuery.compile("application/json", "application/*+json");
-
+    static final Predicate<String> JSON = contentType -> {
+        if(contentType == null) {
+            return false;
+        }
+        // implementation note: manually coded for improved performance
+        if(contentType.startsWith("application/")) {
+            int index = contentType.indexOf(';', 12);
+            if(index != -1) {
+                if(index > 16) {
+                    // application/some+json;charset=utf-8
+                    return contentType.regionMatches(index - 5, "+json", 0, 5);
+                }
+                
+                // application/json;charset=utf-8
+                return contentType.regionMatches(index - 4, "json", 0, 4);
+            } else {
+                // application/json
+                if(contentType.length() == 16) {
+                    return contentType.endsWith("json");
+                }
+                // application/some+json
+                return contentType.endsWith("+json");
+            }
+        }
+        return false;
+    };
 }