Hard-code check for JSON media type for considerable (7x) speedup.

## Description
Hard-code check for JSON media type; avoid use of regexp. Apply previous implementation for unit tests.

Benchmarks run via IDE:
before
JsonMediaTypeBenchmark.contentType1  thrpt    5  10458210,505 ± 112443,545  ops/s
JsonMediaTypeBenchmark.contentType2  thrpt    5   7560926,574 ±  70957,434  ops/s
JsonMediaTypeBenchmark.contentType3  thrpt    5   5107299,369 ±  51905,145  ops/s
JsonMediaTypeBenchmark.contentType4  thrpt    5   3528531,278 ±  34853,242  ops/s
after
JsonMediaTypeBenchmark.contentType1  thrpt    5  78192119,714 ± 4628504,966  ops/s
JsonMediaTypeBenchmark.contentType2  thrpt    5  64037507,380 ± 3227112,459  ops/s
JsonMediaTypeBenchmark.contentType3  thrpt    5  68491591,930 ±  765070,096  ops/s
JsonMediaTypeBenchmark.contentType4  thrpt    5  58140675,375 ±  592186,800  ops/s

## Motivation and Context
JsonMediaType is usually called several times per request/response.

## Types of changes
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.